+++ src/webview/js/incubator.js
@@ -669,3 +669,443 @@ function createGenericEndpointFrom(definition)
 
   return activity
 }
+
+
+
+
+
+// /**
+//  * Implement AABB collision detection for entities with a mesh.
+//  * (https://en.wikipedia.org/wiki/Minimum_bounding_box#Axis-aligned_minimum_bounding_box)
+//  * It sets the specified state on the intersected entities.
+//  *
+//  * @property {string} objects - Selector of the entities to test for collision.
+//  * @property {string} state - State to set on collided entities.
+//  *
+//  */
+// AFRAME.registerComponent('aabb-collider', {
+//     schema: {
+//       objects: {default: ''},
+//       state: {default: 'collided'}
+//     },
+  
+//     init: function () {
+//       this.els = [];
+//       this.collisions = [];
+//       this.elMax = new THREE.Vector3();
+//       this.elMin = new THREE.Vector3();
+//     },
+  
+//     /**
+//      * Update list of entities to test for collision.
+//      */
+//     update: function () {
+//       var data = this.data;
+//       var objectEls;
+  
+//       // Push entities into list of els to intersect.
+//       if (data.objects) {
+//         objectEls = this.el.sceneEl.querySelectorAll(data.objects);
+//       } else {
+//         // If objects not defined, intersect with everything.
+//         objectEls = this.el.sceneEl.children;
+//       }
+//       // Convert from NodeList to Array
+//       this.els = Array.prototype.slice.call(objectEls);
+//     },
+  
+//     tick: (function () {
+//       var boundingBox = new THREE.Box3();
+//       return function () {
+//         var collisions = [];
+//         var el = this.el;
+//         var mesh = el.getObject3D('mesh');
+//         var self = this;
+//         // No mesh, no collisions
+//         if (!mesh) { return; }
+//         // Update the bounding box to account for rotations and
+//         // position changes.
+//         updateBoundingBox();
+//         // Update collisions.
+//         this.els.forEach(intersect);
+//         // Emit events.
+//         collisions.forEach(handleHit);
+//         // No collisions.
+//         if (collisions.length === 0) { self.el.emit('hit', {el: null}); }
+//         // Updated the state of the elements that are not intersected anymore.
+//         this.collisions.filter(function (el) {
+//           return collisions.indexOf(el) === -1;
+//         }).forEach(function removeState (el) {
+//           el.removeState(self.data.state);
+//           el.emit('hitend');
+//         });
+//         // Store new collisions
+//         this.collisions = collisions;
+  
+//         // AABB collision detection
+//         function intersect (el) {
+//           var intersected;
+//           var mesh = el.getObject3D('mesh');
+//           var elMin;
+//           var elMax;
+//           if (!mesh) { return; }
+//           boundingBox.setFromObject(mesh);
+//           elMin = boundingBox.min;
+//           elMax = boundingBox.max;
+//           // Bounding boxes are always aligned with the world coordinate system.
+//           // The collision test checks for the conditions where cubes intersect.
+//           // It's an extension to 3 dimensions of this approach (with the condition negated)
+//           // https://www.youtube.com/watch?v=ghqD3e37R7E
+//           intersected = (self.elMin.x <= elMax.x && self.elMax.x >= elMin.x) &&
+//                         (self.elMin.y <= elMax.y && self.elMax.y >= elMin.y) &&
+//                         (self.elMin.z <= elMax.z && self.elMax.z >= elMin.z);
+//           if (!intersected) { return; }
+//           collisions.push(el);
+//         }
+  
+//         function handleHit (hitEl) {
+//           hitEl.emit('hit');
+//           hitEl.addState(self.data.state);
+//           self.el.emit('hit', {el: hitEl});
+//         }
+  
+//         function updateBoundingBox () {
+//           boundingBox.setFromObject(mesh);
+//           self.elMin.copy(boundingBox.min);
+//           self.elMax.copy(boundingBox.max);
+//         }
+//       };
+//     })()
+//   });
+
+
+
+
+
+
+
+AFRAME.registerComponent('keyboardlistener', {
+    // schema: {
+    //     position: {type: 'string'},
+    //     configMethod: []
+    // },
+    init: function () {
+
+        console.log("init keyboard-listener" + this.el);
+
+        document.addEventListener('keydown', this.notify);
+        document.addEventListener('keyup', this.notify);
+    },
+
+    notify: function(event) {
+        if(event.code == "ShiftRight" || event.code == "ShiftLeft")
+        {
+            console.log("shift: "+event.type+"/"+event.key)
+
+            let pressed = (event.type == "keydown")
+
+            this.querySelectorAll('[detachable]').forEach(function(activity){
+                activity.components.detachable.keypressed(pressed) // .setAttribute('position', '0 0 0');
+            });
+        }   
+    }
+
+    // update: function () {},
+    // tick: function () {},
+    // remove: function () {},
+    // pause: function () {},
+    // play: function () {}
+});
+
+
+
+
+//=========================================
+
+
+AFRAME.registerComponent('detachable', {
+    // schema: {
+    //     position: {type: 'string'},
+    //     configMethod: []
+    // },
+    init: function () {
+        console.log("init attachable" + this.el);
+
+        this.shiftPressed = false
+
+        this.colorDefault = this.el.getAttribute('color') || this.el.getAttribute('material').color
+        this.colorHighlight = "yellow"
+
+        this.opacityDefault  = this.el.getAttribute('opacity')|| this.el.getAttribute('material').opacity
+        this.opacityDetached = 0.3
+    },
+
+    keypressed: function(pressed) {
+
+        //keep shift status
+        this.shiftPressed = pressed
+
+        if(pressed && !this.isLastInBranch())
+        {
+            this.el.setAttribute('color', this.colorHighlight)
+        }
+        else
+        {
+            if(this.colorDefault)
+                this.el.setAttribute('color', this.colorDefault)
+            else   
+                this.el.removeAttribute('color')
+        }
+    },
+
+    isLastInBranch: function() {
+        //get activity that follows
+        let next = getNextActivity(this.el);
+        let previous = getPreviousActivity(this.el)
+
+        //not allowed to detach last activity in choice branch
+        if(previous.getAttribute('processor-type') == 'choice-start' && next.getAttribute('processor-type') == 'choice-end')
+        {
+            return true
+        }
+
+        return false
+    },
+
+    detach: function () {
+
+        //ignore if already detached
+        if(this.detached)
+        {
+            return
+        }
+
+        //helper
+        let activity = this.el
+
+        //not allowed to detach last activity in choice branch
+        if(this.isLastInBranch())
+        {
+            return
+        }
+
+        //do not allow new activies while moving objects
+        enableToButtons(false)
+
+        //flag
+        this.detached = true
+
+        //make all activities static
+        for(entity of activity.parentNode.children) {
+            if(entity != activity)
+                entity.setAttribute('class', 'not-clickable')
+        }
+
+        //make look dimmer
+        activity.setAttribute('opacity', this.opacityDetached)
+
+        //add collision detection and event listenerss
+        activity.setAttribute('aabb-collider', 'objects: a-cylinder')
+        activity.addEventListener('hitclosest',      this.candidateHighlight);
+        activity.addEventListener('hitclosestclear', this.candidateClear);
+
+
+        let linkForward = null
+
+        //get activity that follows
+        let next = getNextActivity(activity);
+
+        //if this is the last process activity
+        if(next == null)
+        {
+            let previous = getPreviousActivity(activity);
+
+            disposableLink = detachForwardLink(previous)
+
+            detachBackwardsLink(activity)
+
+            //delete backwards link
+            disposableLink.parentNode.removeChild(disposableLink);
+
+            //add listener to reattach
+            activity.addEventListener('mouseup', this.reattach)
+
+            syncEditor()
+            return
+        }
+
+
+        if(next.getAttribute('processor-type') == 'choice-end')
+        {
+            linkForward = getForwardLink(activity)
+        }
+
+        //get links
+        linkForward = detachBackwardsLink(next, linkForward);
+        let linkBackwards = detachBackwardsLink(activity)
+
+
+        //rewire
+        if(linkBackwards)
+        {
+            attachDestinationToLink(next, linkBackwards);
+        }
+
+        if(activity.localName == "a-box")
+        {
+            //obtain end
+            let boxEnd = activity.querySelector('[processor-type=multicast-end]')
+
+            //obtain all link references the activity has
+            refLinks = JSON.parse(boxEnd.getAttribute("links"));
+
+            //we remove the link (detach) from the activity
+            refLinks.splice(refLinks.indexOf(linkForward.id),1)[0];
+
+            //we update the activity (with link stripped out)
+            boxEnd.setAttribute("links", JSON.stringify(refLinks))
+        }
+        else
+        {
+            //detached activity would be orphan with no links
+            activity.setAttribute('links', "[]")
+        }
+
+        //delete backwards link
+        linkForward.parentNode.removeChild(linkForward);
+
+        //add listener to reattach
+        activity.addEventListener('mouseup', this.reattach)
+
+        syncEditor()
+    },
+
+    candidateHighlight: function (event) {
+        // console.log('target cylinder: ' + event.detail.el)
+
+        //if one previously existed, we reset color
+        if(this.components.detachable.linkDetected)
+        {
+            this.components.detachable.linkDetected.removeAttribute('color')
+        }
+
+        event.detail.el.setAttribute('color', 'yellow')
+        this.setAttribute('color', 'yellow')
+
+        //candidate link in proximity
+        this.components.detachable.linkDetected = event.detail.el
+
+        console.log('candidateHighlight: ' + this.components.detachable.linkDetected.id)
+
+    },
+    candidateClear: function (event) {
+        // console.log('target cylinder: ' + event.detail.el)
+        event.detail.el.removeAttribute('color')
+        this.setAttribute('color', '#ACF5F5')
+
+        // clear candidate
+        this.components.detachable.linkDetected = null
+    },
+
+    reattach: function ()
+    {
+        let detachable = this.components.detachable
+
+        //ignore if not detached
+        if(!detachable.detached)
+        {
+            return
+        }
+
+        //helper
+        let activity = this
+    
+        //if candidate exists
+        if(detachable.linkDetected)
+        {
+            //reactivate all activities
+            for(entity of activity.parentNode.children) {
+                if(entity != activity)
+                    entity.setAttribute('class', 'clickable')
+            }
+
+            //allow again new activies
+            enableToButtons(true)
+
+            //remove listeners
+            activity.removeEventListener('mouseup', detachable.reattach);    
+            activity.removeEventListener('hitclosest', detachable.joinHighlight);
+            activity.removeEventListener('hitclosestclear', detachable.joinClear);
+
+            //remove collision detectioon
+            activity.removeAttribute('aabb-collider')
+        
+            //obtain necessray artifacts
+            let link = detachable.linkDetected
+            let source = document.getElementById(link.getAttribute('source'))
+    
+    
+            // let copy = activity.cloneNode(true)
+            // activity.parentNode.removeChild(activity)
+            // source.parentNode.appendChild(copy)
+    
+            //restore material attributes
+            activity.setAttribute('opacity', detachable.opacityDefault)
+            activity.setAttribute('color', detachable.colorDefault)
+            link.removeAttribute('color')
+
+            let newLink
+
+            //obtain endpoint of link
+            let destination = activity.parentNode.querySelector('#'+link.getAttribute('destination'))
+
+            //when reattaching a box (multicast)
+            if(activity.localName == "a-box")
+            {
+                let boxStart = activity.querySelector('[processor-type=multicast-start]')
+                let boxEnd   = activity.querySelector('[processor-type=multicast-end]')
+
+                //we keep link reference.
+                let reusableLink = detachBackwardsLink(destination, link)
+
+                //create new link
+                newLink = createLink(boxEnd, destination, false, false)
+
+                //attach end of link to start of box
+                attachDestinationToLink(boxStart, reusableLink)
+            }
+            else
+            {
+                //we keep link reference.
+                let reusableLink = detachBackwardsLink(destination, link)
+
+                //create new link
+                newLink = createLink(activity, destination, false, false)
+
+                //attach end of link to activity
+                attachDestinationToLink(activity, reusableLink)
+            }
+
+            //add link to scene
+            activity.parentNode.appendChild(newLink);
+    
+            // let copy = activity.cloneNode(true)
+            // activity.parentNode.removeChild(activity)
+            // source.parentNode.appendChild(copy)
+            // activity.after(source);
+            // insertActivity(activity, source)
+
+            //reset variables
+            detachable.linkDetected = null
+            detachable.detached = false
+
+            syncEditor()
+        }
+    }
+
+    // update: function () {},
+    // tick: function () {},
+    // remove: function () {},
+    // pause: function () {},
+    // play: function () {}
+});